+++ pkg/providers/dns/provider.go
@@ -75,5 +75,7 @@ func (p *Subprovider) FrameworkResources() []func() resource.Resource {
 
 // FrameworkDataSources returns the DNS data sources implemented using terraform-plugin-framework
 func (p *Subprovider) FrameworkDataSources() []func() datasource.DataSource {
-	return []func() datasource.DataSource{}
+	return []func() datasource.DataSource{
+		NewZoneDNSSecStatusDataSource,
+	}
 }