+++ pkg/providers/dns/testdata/TestDataZoneDnsSecStatus/missing_zone_name.tf
@@ -1,0 +1,6 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_zone_dnssec_status" "test" {
+}